+++ src/realm/group.cpp
@@ -1652,6 +1652,10 @@ public:
     {
         return true; // No-op
     }
+    bool dictionary_set(size_t, Mixed)
+    {
+        return true; // No-op
+    }
     bool dictionary_erase(size_t, Mixed)
     {
         return true; // No-op