+++ src/realm/object-store/impl/transact_log_handler.cpp
@@ -291,6 +291,10 @@ public:
     {
         return true;
     }
+    bool dictionary_set(size_t, Mixed)
+    {
+        return true;
+    }
     bool dictionary_erase(size_t, Mixed)
     {
         return true;
@@ -453,6 +457,13 @@ public:
     {
         if (m_active_collection) {
             m_active_collection->insert(index);
+        }
+        return true;
+    }
+    bool dictionary_set(size_t index, Mixed)
+    {
+        if (m_active_collection) {
+            m_active_collection->modify(index);
         }
         return true;
     }